requirements.txt
cycler==0.10.0
face-recognition-models==0.3.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.14.0
